# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Webjal Demo Studio API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface PackageItem {
  name: string;
  price: string;
  features: string[];
}

export interface Project {
  id: number;
  businessName: string;
  category: string;
  /** @nullable */
  tagline?: string | null;
  /** @nullable */
  about?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  whatsapp?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  googleMapsLink?: string | null;
  /** @nullable */
  instagramLink?: string | null;
  services: string[];
  packages: PackageItem[];
  /** @nullable */
  ctaText?: string | null;
  /** @nullable */
  primaryColor?: string | null;
  /** @nullable */
  secondaryColor?: string | null;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  heroImageUrl?: string | null;
  galleryImages: string[];
  /** @nullable */
  templateId?: number | null;
  /** @nullable */
  generatedHtml?: string | null;
  status: string;
  createdAt: string;
  updatedAt: string;
}

export interface CategoryCount {
  category: string;
  count: number;
}

export interface DashboardStats {
  totalTemplates: number;
  totalProjects: number;
  recentProjects: Project[];
  templatesByCategory: CategoryCount[];
  projectsByCategory: CategoryCount[];
}

export interface Template {
  id: number;
  name: string;
  category: string;
  /** @nullable */
  description?: string | null;
  htmlContent: string;
  /** @nullable */
  cssContent?: string | null;
  /** @nullable */
  jsContent?: string | null;
  /** @nullable */
  thumbnailUrl?: string | null;
  placeholders: string[];
  createdAt: string;
  updatedAt: string;
}

export interface TemplateInput {
  name: string;
  category: string;
  description?: string;
  htmlContent: string;
  cssContent?: string;
  jsContent?: string;
  thumbnailUrl?: string;
  placeholders?: string[];
}

export interface TemplateUpdate {
  name?: string;
  category?: string;
  description?: string;
  htmlContent?: string;
  cssContent?: string;
  jsContent?: string;
  thumbnailUrl?: string;
  placeholders?: string[];
}

export interface ProjectInput {
  businessName: string;
  category: string;
  tagline?: string;
  about?: string;
  phone?: string;
  whatsapp?: string;
  email?: string;
  address?: string;
  googleMapsLink?: string;
  instagramLink?: string;
  services?: string[];
  packages?: PackageItem[];
  ctaText?: string;
  primaryColor?: string;
  secondaryColor?: string;
  logoUrl?: string;
  heroImageUrl?: string;
  galleryImages?: string[];
  templateId?: number;
}

export interface ProjectUpdate {
  businessName?: string;
  category?: string;
  tagline?: string;
  about?: string;
  phone?: string;
  whatsapp?: string;
  email?: string;
  address?: string;
  googleMapsLink?: string;
  instagramLink?: string;
  services?: string[];
  packages?: PackageItem[];
  ctaText?: string;
  primaryColor?: string;
  secondaryColor?: string;
  logoUrl?: string;
  heroImageUrl?: string;
  galleryImages?: string[];
  templateId?: number;
  generatedHtml?: string;
  status?: string;
}

export type ListTemplatesParams = {
category?: string;
};

export type ListProjectsParams = {
category?: string;
search?: string;
};