# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Webjal Demo Studio API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: templates
    description: Website template management
  - name: projects
    description: Client demo project management
  - name: dashboard
    description: Dashboard statistics
  - name: uploads
    description: File upload management

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard statistics
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

  /templates:
    get:
      operationId: listTemplates
      tags: [templates]
      summary: List all templates
      parameters:
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Template"
    post:
      operationId: createTemplate
      tags: [templates]
      summary: Create a new template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TemplateInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"

  /templates/{id}:
    get:
      operationId: getTemplate
      tags: [templates]
      summary: Get a template by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateTemplate
      tags: [templates]
      summary: Update a template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TemplateUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
    delete:
      operationId: deleteTemplate
      tags: [templates]
      summary: Delete a template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all client projects
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new client project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /projects/{id}/generate:
    post:
      operationId: generateProject
      tags: [projects]
      summary: Generate website HTML from template + client data
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

  /projects/{id}/duplicate:
    post:
      operationId: duplicateProject
      tags: [projects]
      summary: Duplicate an existing project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Duplicated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"


components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]

    DashboardStats:
      type: object
      properties:
        totalTemplates:
          type: integer
        totalProjects:
          type: integer
        recentProjects:
          type: array
          items:
            $ref: "#/components/schemas/Project"
        templatesByCategory:
          type: array
          items:
            $ref: "#/components/schemas/CategoryCount"
        projectsByCategory:
          type: array
          items:
            $ref: "#/components/schemas/CategoryCount"
      required: [totalTemplates, totalProjects, recentProjects, templatesByCategory, projectsByCategory]

    CategoryCount:
      type: object
      properties:
        category:
          type: string
        count:
          type: integer
      required: [category, count]

    Template:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        description:
          type: ["string", "null"]
        htmlContent:
          type: string
        cssContent:
          type: ["string", "null"]
        jsContent:
          type: ["string", "null"]
        thumbnailUrl:
          type: ["string", "null"]
        placeholders:
          type: array
          items:
            type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, category, htmlContent, placeholders, createdAt, updatedAt]

    TemplateInput:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        description:
          type: string
        htmlContent:
          type: string
        cssContent:
          type: string
        jsContent:
          type: string
        thumbnailUrl:
          type: string
        placeholders:
          type: array
          items:
            type: string
      required: [name, category, htmlContent]

    TemplateUpdate:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        description:
          type: string
        htmlContent:
          type: string
        cssContent:
          type: string
        jsContent:
          type: string
        thumbnailUrl:
          type: string
        placeholders:
          type: array
          items:
            type: string

    Project:
      type: object
      properties:
        id:
          type: integer
        businessName:
          type: string
        category:
          type: string
        tagline:
          type: ["string", "null"]
        about:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        whatsapp:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        googleMapsLink:
          type: ["string", "null"]
        instagramLink:
          type: ["string", "null"]
        services:
          type: array
          items:
            type: string
        packages:
          type: array
          items:
            $ref: "#/components/schemas/PackageItem"
        ctaText:
          type: ["string", "null"]
        primaryColor:
          type: ["string", "null"]
        secondaryColor:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        heroImageUrl:
          type: ["string", "null"]
        galleryImages:
          type: array
          items:
            type: string
        templateId:
          type: ["integer", "null"]
        generatedHtml:
          type: ["string", "null"]
        status:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, businessName, category, services, packages, galleryImages, status, createdAt, updatedAt]

    PackageItem:
      type: object
      properties:
        name:
          type: string
        price:
          type: string
        features:
          type: array
          items:
            type: string
      required: [name, price, features]

    ProjectInput:
      type: object
      properties:
        businessName:
          type: string
        category:
          type: string
        tagline:
          type: string
        about:
          type: string
        phone:
          type: string
        whatsapp:
          type: string
        email:
          type: string
        address:
          type: string
        googleMapsLink:
          type: string
        instagramLink:
          type: string
        services:
          type: array
          items:
            type: string
        packages:
          type: array
          items:
            $ref: "#/components/schemas/PackageItem"
        ctaText:
          type: string
        primaryColor:
          type: string
        secondaryColor:
          type: string
        logoUrl:
          type: string
        heroImageUrl:
          type: string
        galleryImages:
          type: array
          items:
            type: string
        templateId:
          type: integer
      required: [businessName, category]

    ProjectUpdate:
      type: object
      properties:
        businessName:
          type: string
        category:
          type: string
        tagline:
          type: string
        about:
          type: string
        phone:
          type: string
        whatsapp:
          type: string
        email:
          type: string
        address:
          type: string
        googleMapsLink:
          type: string
        instagramLink:
          type: string
        services:
          type: array
          items:
            type: string
        packages:
          type: array
          items:
            $ref: "#/components/schemas/PackageItem"
        ctaText:
          type: string
        primaryColor:
          type: string
        secondaryColor:
          type: string
        logoUrl:
          type: string
        heroImageUrl:
          type: string
        galleryImages:
          type: array
          items:
            type: string
        templateId:
          type: integer
        generatedHtml:
          type: string
        status:
          type: string