# lib/api-zod/src/generated/types/categoryCount.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Webjal Demo Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface CategoryCount {
  category: string;
  count: number;
}