# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Webjal Demo Studio API
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryCount } from './categoryCount';
import type { Project } from './project';

export interface DashboardStats {
  totalTemplates: number;
  totalProjects: number;
  recentProjects: Project[];
  templatesByCategory: CategoryCount[];
  projectsByCategory: CategoryCount[];
}