# lib/api-zod/src/generated/types/packageItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Webjal Demo Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface PackageItem {
  name: string;
  price: string;
  features: string[];
}