# lib/api-zod/src/generated/types/projectInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Webjal Demo Studio API
 * OpenAPI spec version: 0.1.0
 */
import type { PackageItem } from './packageItem';

export interface ProjectInput {
  businessName: string;
  category: string;
  tagline?: string;
  about?: string;
  phone?: string;
  whatsapp?: string;
  email?: string;
  address?: string;
  googleMapsLink?: string;
  instagramLink?: string;
  services?: string[];
  packages?: PackageItem[];
  ctaText?: string;
  primaryColor?: string;
  secondaryColor?: string;
  logoUrl?: string;
  heroImageUrl?: string;
  galleryImages?: string[];
  templateId?: number;
}