# lib/api-zod/src/generated/types/templateInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Webjal Demo Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface TemplateInput {
  name: string;
  category: string;
  description?: string;
  htmlContent: string;
  cssContent?: string;
  jsContent?: string;
  thumbnailUrl?: string;
  placeholders?: string[];
}